+++ happybase/table.py
@@ -103,12 +103,12 @@ class Table(object):
 
         if timestamp is None:
             rows = self.connection.client.getRowWithColumns(
-                self.name, row, columns)
+                self.name, row, columns, {})
         else:
             if not isinstance(timestamp, Integral):
                 raise TypeError("'timestamp' must be an integer")
             rows = self.connection.client.getRowWithColumnsTs(
-                self.name, row, columns, timestamp)
+                self.name, row, columns, timestamp, {})
 
         if not rows:
             return {}
@@ -143,7 +143,7 @@ class Table(object):
 
         if timestamp is None:
             results = self.connection.client.getRowsWithColumns(
-                self.name, rows, columns)
+                self.name, rows, columns, {})
         else:
             if not isinstance(timestamp, Integral):
                 raise TypeError("'timestamp' must be an integer")
@@ -155,7 +155,7 @@ class Table(object):
                 columns = self._column_family_names()
 
             results = self.connection.client.getRowsWithColumnsTs(
-                self.name, rows, columns, timestamp)
+                self.name, rows, columns, timestamp, {})
 
         return [(r.row, make_row(r.columns, include_timestamp))
                 for r in results]
@@ -191,12 +191,12 @@ class Table(object):
 
         if timestamp is None:
             cells = self.connection.client.getVer(
-                self.name, row, column, versions)
+                self.name, row, column, versions, {})
         else:
             if not isinstance(timestamp, Integral):
                 raise TypeError("'timestamp' must be an integer")
             cells = self.connection.client.getVerTs(
-                self.name, row, column, timestamp, versions)
+                self.name, row, column, timestamp, versions, {})
 
         if include_timestamp:
             return map(make_cell_timestamp, cells)
@@ -286,17 +286,17 @@ class Table(object):
             if row_stop is None:
                 if timestamp is None:
                     scan_id = self.connection.client.scannerOpen(
-                        self.name, row_start, columns)
+                        self.name, row_start, columns, {})
                 else:
                     scan_id = self.connection.client.scannerOpenTs(
-                        self.name, row_start, columns, timestamp)
+                        self.name, row_start, columns, timestamp, {})
             else:
                 if timestamp is None:
                     scan_id = self.connection.client.scannerOpenWithStop(
-                        self.name, row_start, row_stop, columns)
+                        self.name, row_start, row_stop, columns, {})
                 else:
                     scan_id = self.connection.client.scannerOpenWithStopTs(
-                        self.name, row_start, row_stop, columns, timestamp)
+                        self.name, row_start, row_stop, columns, timestamp, {})
 
         else:
             # The scan's caching size is set to the batch_size, so that
@@ -310,9 +310,10 @@ class Table(object):
                 columns=columns,
                 caching=batch_size,
                 filterString=filter,
+                batchSize=batch_size,
             )
             scan_id = self.connection.client.scannerOpenWithScan(
-                self.name, scan)
+                self.name, scan, {})
 
         logger.debug("Opened scanner (id=%d) on '%s'", scan_id, self.name)
 
@@ -350,26 +351,30 @@ class Table(object):
     # Data manipulation
     #
 
-    def put(self, row, data, timestamp=None):
+    def put(self, row, data, timestamp=None, wal=True):
         """Store data in the table.
 
         This method stores the data in the `data` argument for the row
         specified by `row`. The `data` argument is dictionary that maps columns
         to values. Column names must include a family and qualifier part, e.g.
         `cf:col`, though the qualifier part may be the empty string, e.g.
-        `cf:`. The `timestamp` argument is optional.
+        `cf:`.
 
         Note that, in many situations, :py:meth:`batch()` is a more appropriate
         method to manipulate data.
 
+        .. versionadded:: 0.7
+           `wal` parameter
+
         :param str row: the row key
         :param dict data: the data to store
         :param int timestamp: timestamp (optional)
+        :param wal bool: whether to write to the WAL (optional)
         """
-        with self.batch(timestamp=timestamp) as batch:
+        with self.batch(timestamp=timestamp, wal=wal) as batch:
             batch.put(row, data)
 
-    def delete(self, row, columns=None, timestamp=None):
+    def delete(self, row, columns=None, timestamp=None, wal=True):
         """Delete data from the table.
 
         This method deletes all columns for the row specified by `row`, or only
@@ -378,21 +383,19 @@ class Table(object):
         Note that, in many situations, :py:meth:`batch()` is a more appropriate
         method to manipulate data.
 
+        .. versionadded:: 0.7
+           `wal` parameter
+
         :param str row: the row key
         :param list_or_tuple columns: list of columns (optional)
         :param int timestamp: timestamp (optional)
+        :param wal bool: whether to write to the WAL (optional)
         """
-        if columns is None:
-            if timestamp is None:
-                self.connection.client.deleteAllRow(self.name, row)
-            else:
-                self.connection.client.deleteAllRowTs(
-                    self.name, row, timestamp)
-        else:
-            with self.batch(timestamp=timestamp) as batch:
-                batch.delete(row, columns)
+        with self.batch(timestamp=timestamp, wal=wal) as batch:
+            batch.delete(row, columns)
 
-    def batch(self, timestamp=None, batch_size=None, transaction=False):
+    def batch(self, timestamp=None, batch_size=None, transaction=False,
+              wal=True):
         """Create a new batch operation for this table.
 
         This method returns a new :py:class:`Batch` instance that can be used
@@ -408,11 +411,23 @@ class Table(object):
         used as context manager in a ``with`` block of code. The `transaction`
         flag cannot be used in combination with `batch_size`.
 
+        The `wal` argument determines whether mutations should be
+        written to the HBase Write Ahead Log (WAL). This flag can only
+        be used with recent HBase versions. If specified, it provides
+        a default for all the put and delete operations on this batch.
+        This default value can be overridden for individual operations
+        using the `wal` argument to :py:meth:`Batch.put` and
+        :py:meth:`Batch.delete`.
+
+        .. versionadded:: 0.7
+           `wal` parameter
+
         :param bool transaction: whether this batch should behave like
                                  a transaction (only useful when used as a
                                  context manager)
         :param int batch_size: batch size (optional)
         :param int timestamp: timestamp (optional)
+        :param wal bool: whether to write to the WAL (optional)
 
         :return: Batch instance
         :rtype: :py:class:`Batch`